[Brijesh] ADD. Implemented buy() method

+++ OOPS/src/main/java/com/bridgelabz/Services/IStockAccount.java
@@ -9,8 +9,8 @@ package com.bridgelabz.Services;
 public interface IStockAccount 
 {
 	double valueOF();
-	void buy(int amount,String symbol);
-	void sell(int amount,String symbol);
+	void buy(long amount,String symbol);
+	void sell(long amount,String symbol);
 	void save(String fileName);
 	void printReport();
 }
